Rename package ui.event to ui.events
