+++ app/src/main/java/com/fake/soundremote/ui/events/EventsNavigation.kt
@@ -1,4 +1,4 @@
-package com.fake.soundremote.ui.event
+package com.fake.soundremote.ui.events
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
